feat(module_resolution): rename resolvers to resolver
