+++ src/front/mergers/package/module_resolution.rs
@@ -1,3 +1,3 @@
 pub mod merged_module;
 pub mod module_merger;
-mod resolvers;
+mod resolver;
