+++ src/front/mergers/package/module_resolution/resolver.rs
